test(dw): split dataset tests into multiple files to improve readability
